fix underscore to hyphen for package name
I remember making this PR days ago, but it didn't seem to go through... Apologies if I made and you decided to nuke it completely

pip does some magic to convert underscore package names to hyphen package names internally. Underscores are discouraged for module names by python (and not possible on pypi afaik). Furthermore, conda is not as magic as pip and will not find the appropriate package that way.

From pep8:
> Modules should have short, all-lowercase names. Underscores can be used in the module name if it improves readability. Python packages should also have short, all-lowercase names, although the use of underscores is discouraged.